Update for testnet validator start script

+++ loadScript/deployValidator.sh
@@ -15,7 +15,11 @@ read -s PASSWD
 echo 'Confirm your password of validator account:'
 read -s PASSWD2
 echo ''
-DOCKERIMG=wanchain/client-go:2.1.8
+
+echo ''
+read -p "Do you want save your password to disk for auto restart? (N/y): " savepasswd
+
+DOCKERIMG=wanchain/client-go:2.2.1-beta.1
 NETWORK=--testnet
 NETWORKPATH=testnet
 
@@ -74,7 +78,11 @@ echo 'Please wait a few seconds...'
 
 sleep 5
 
-sudo rm ~/.wanchain/pw.txt
+if [ "$savepasswd" == "Y" ] || [ "$savepasswd" == "y" ]; then
+    sudo docker container update --restart=always gwan
+else
+    sudo rm ~/.wanchain/pw.txt
+fi
 
 KEYSTOREFILE=$(sudo ls ~/.wanchain/testnet/keystore/)
